refactor(admin): move danger actions to own card
